Rename io.h to io_shaderc.h
This makes libshaderc possible to compile on Windows under Bazel, due to https://github.com/bazelbuild/bazel/issues/11672.
