+++ glslc/src/dependency_info.cc
@@ -19,7 +19,7 @@
 #include <sstream>
 
 #include "file.h"
-#include "libshaderc_util/io.h"
+#include "libshaderc_util/io_shaderc.h"
 
 namespace glslc {
