+++ libshaderc_util/src/io.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "libshaderc_util/io.h"
+#include "libshaderc_util/io_shaderc.h"
 
 #include "libshaderc_util/universal_unistd.h"
